commented out tensorboard

+++ text-seg/gen_rlf_data.py
@@ -11,7 +11,7 @@ if not os.path.exists(output_dir):
     os.makedirs(output_dir)
 
 sections = traverse_json_dir(json_dir, toke_to_sent=False, limit_paragraphs=0)
-l_tr = float(len(sections))*0.8
+l_tr = int(float(len(sections))*0.8)
 
 with open(os.path.join(output_dir,"train.txt"),'w') as fout:
     fout.write("\n".join(sections[:l_tr])+"\n")
